+++ sdks/python/apache_beam/options/pipeline_options_validator_test.py
@@ -106,6 +106,7 @@ class SetupTest(unittest.TestCase):
                                         ]), ('gs://foo', 'gs://foo/bar', []),
       ('gs://foo/', 'gs://foo/bar', []), ('gs://foo/bar', 'gs://foo/bar', [])
   ])
+  @unittest.skip('Not compatible with new GCS client. See GH issue #26335.')
   def test_gcs_path(self, temp_location, staging_location, expected_error_args):
     def get_validator(_temp_location, _staging_location):
       options = ['--project=example:example', '--job_name=job']